pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sql-agent-gemini"
version = "0.1.0"
description = "Script-based SQL Agent with Gemini and Streamlit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "google-genai>=1.30.0",
  "streamlit>=1.40.0",
  "pypyodbc>=1.3.6",
  "pandas>=2.2.0",
  "python-dotenv>=1.0.1"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
